runtime/api/requirements.txt
fastapi==0.115.12
starlette==0.46.2
uvicorn==0.34.1
python-multipart==0.0.20
pandas==2.2.3
pyarrow==19.0.1
mcp[cli]==1.27.0
databricks-sdk>=0.50.0
